ep.26

+++ app/Http/Controllers/UserController.php
@@ -14,7 +14,7 @@ class UserController extends Controller
     public function show(User $user)
     {
         $ideas = $user->ideas()->paginate(5);
-        return view('user.show',compact('user','ideas'));
+        return view('user.show', compact('user', 'ideas'));
     }
 
     /**
@@ -22,9 +22,9 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        $editing = true;
+        //$editing = true;
         $ideas = $user->ideas()->paginate(5);
-        return view('user.edit',compact('user','editing','ideas'));
+        return view('user.edit', compact('user',  'ideas'));
     }
 
     /**
@@ -38,15 +38,17 @@ class UserController extends Controller
     public function update(User $user)
     {
         $data = request()->validate([
-            'image'=>'image',
-            'name'=>'required|min:3|max:40',
-            'bio'=>'sometimes|nullable|min:3|max:255',
+            'image' => 'image',
+            'name' => 'required|min:3|max:40',
+            'bio' => 'sometimes|nullable|min:3|max:255',
         ]);
 
-        if(request('image')){
-            $imagePath = request('image')->store('profile','public');
+        if (request('image')) {
+            $imagePath = request('image')->store('profile', 'public');
             $data['image'] = $imagePath;
-            Storage::disk('public')->delete($user->image);
+            if ($user->image) {
+                Storage::disk('public')->delete($user->image);
+            }
         }
 
         $user->update($data);
